+++ src/DuetControlServer/Program.cs
@@ -9,7 +9,6 @@ using System.Linq;
 using System.Net.Sockets;
 using System.Reflection;
 using System.Runtime.Loader;
-using System.Text;
 using System.Text.Json;
 using System.Threading;
 using System.Threading.Tasks;